fixed: Incorrect conversion between integer types
Messages thrown by Code scanning was issuing a code mistake happens.

Since I am a go beginner, I used Notepad++ to fix these issues.

It may look different than https://github.com/miekg/dns/pull/1212

+++ generate_test.go
@@ -162,7 +162,7 @@ func TestGenerateModToPrintf(t *testing.T) {
 	tests := []struct {
 		mod        string
 		wantFmt    string
-		wantOffset int64
+		wantOffset int
 		wantErr    bool
 	}{
 		{"0,0,d", "%d", 0, false},
